test_ViT/requirements.txt
torch
torchvision
transformers
Pillow
numpy
tkinter
matplotlib
